requirements.txt
flask==2.2.2
flask-sqlalchemy==3.0.2
python-dotenv
psycopg2-binary
Flask-Script==2.0.6
Flask-Migrate==2.5.2
Pillow
PyJWT==2.1.0
Werkzeug